refactor(clipboard): move paste read pipeline and paste bytes into action module
